unified merge() and dolt_merge(), updated BATS to stage changes before committing, some sql-server tests still failing
