+++ go/libraries/doltcore/sqle/dfunctions/dolt_commit.go
@@ -41,10 +41,10 @@ func NewDoltCommitFunc(args ...sql.Expression) (sql.Expression, error) {
 }
 
 func (d DoltCommitFunc) Eval(ctx *sql.Context, row sql.Row) (interface{}, error) {
-	return makeDoltCommit(ctx, row, d.Children())
+	return doDoltCommit(ctx, row, d.Children())
 }
 
-func makeDoltCommit(ctx *sql.Context, row sql.Row, exprs []sql.Expression) (interface{}, error) {
+func doDoltCommit(ctx *sql.Context, row sql.Row, exprs []sql.Expression) (interface{}, error) {
 	// Get the information for the sql context.
 	dbName := ctx.GetCurrentDatabase()
 	ap := cli.CreateCommitArgParser()
